refactor: Restructure part domain folder and test runner setup
