+++ backend/src/models/mod.rs
@@ -1,2 +1,2 @@
-pub mod part;
+// pub mod part;
 pub mod user;
