+++ backend/src/part/mod.rs
@@ -1,0 +1,3 @@
+pub mod domain;
+pub mod route;
+pub mod service;
