+++ backend/src/part/service/delete.rs
