+++ tests/api/run_all.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+set -e
+
+./tests/wait-for-backend.sh
+./tests/part/api_test.sh
